(PC-17816) refactor(profile): extract and move all functions in utils to helpers
