+++ src/features/bookOffer/components/BookingDetails.tsx
@@ -7,7 +7,7 @@ import { isApiError } from 'api/apiHelpers'
 import { OfferStockResponse } from 'api/gen'
 import { DuoChoiceSelector } from 'features/bookOffer/components/DuoChoiceSelector'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
-import { useIsUserUnderage } from 'features/profile/utils'
+import { useIsUserUnderage } from 'features/profile/helpers/useIsUserUnderage'
 import { useLogOfferConversion } from 'libs/algolia/analytics/logOfferConversion'
 import { campaignTracker, CampaignEvents } from 'libs/campaign'
 import { analytics } from 'libs/firebase/analytics'
