+++ src/features/bookings/components/BookingDetailsCancelButton.native.test.tsx
@@ -8,7 +8,7 @@ import {
 } from 'features/bookings/components/BookingDetailsCancelButton'
 import { bookingsSnap } from 'features/bookings/fixtures/bookingsSnap'
 import { Booking } from 'features/bookings/types'
-import { isUserExBeneficiary } from 'features/profile/utils'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { fireEvent, render } from 'tests/utils'
 
 mockdate.set(new Date('2020-12-01T00:00:00Z'))
