+++ src/features/bookings/components/BookingDetailsCancelButton.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { getBookingProperties } from 'features/bookings/helpers'
 import { Booking } from 'features/bookings/types'
-import { isUserExBeneficiary } from 'features/profile/utils'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { formatToCompleteFrenchDate } from 'libs/parsers'
 import { useSubcategory } from 'libs/subcategories'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
