+++ src/features/bookings/components/CancelBookingModal.tsx
@@ -9,7 +9,8 @@ import { useCancelBookingMutation } from 'features/bookings/api'
 import { Booking } from 'features/bookings/types'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
-import { isUserBeneficiary, isUserExBeneficiary } from 'features/profile/utils'
+import { isUserBeneficiary } from 'features/profile/helpers/isUserBeneficiary'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { analytics } from 'libs/firebase/analytics'
 import { useNetInfoContext } from 'libs/network/NetInfoWrapper'
 import { formatToFrenchDecimal } from 'libs/parsers'
