+++ src/features/favorites/components/Buttons/BookingButton.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 
 import { FavoriteOfferResponse, UserProfileResponse } from 'api/gen'
 import { hasEnoughCredit } from 'features/offer/helpers/useHasEnoughCredit/useHasEnoughCredit'
-import { isUserBeneficiary, isUserExBeneficiary } from 'features/profile/utils'
+import { isUserBeneficiary } from 'features/profile/helpers/isUserBeneficiary'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ExternalTouchableLink } from 'ui/components/touchableLink/ExternalTouchableLink'
 import { ExternalSite } from 'ui/svg/icons/ExternalSite'
