+++ src/features/offer/helpers/useCtaWordingAndAction/useCtaWordingAndAction.ts
@@ -9,7 +9,7 @@ import {
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useEndedBookingFromOfferId } from 'features/bookings/api'
 import { OfferModal } from 'features/offer/enums'
-import { isUserUnderageBeneficiary } from 'features/profile/utils'
+import { isUserUnderageBeneficiary } from 'features/profile/helpers/isUserUnderageBeneficiary'
 import { analytics } from 'libs/firebase/analytics'
 import { useSubcategoriesMapping } from 'libs/subcategories'
 import { Subcategory } from 'libs/subcategories/types'
