+++ src/features/profile/components/Badges/ProfileBadge.native.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { ProfileBadge } from 'features/profile/components/Badges/ProfileBadge'
-import { shouldOpenInbox } from 'features/profile/utils'
+import { shouldOpenInbox } from 'features/profile/helpers/shouldOpenInbox'
 import { render, fireEvent } from 'tests/utils'
 import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 import { EmailFilled } from 'ui/svg/icons/EmailFilled'
