+++ src/features/profile/components/Badges/SubscriptionMessageBadge.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 import { SubscriptionMessage } from 'api/gen'
 import { ProfileBadge } from 'features/profile/components/Badges/ProfileBadge'
 import { Subtitle } from 'features/profile/components/Subtitle/Subtitle'
-import { matchSubscriptionMessageIconToSvg } from 'features/profile/utils'
+import { matchSubscriptionMessageIconToSvg } from 'features/profile/helpers/matchSubscriptionMessageIconToSvg'
 import { formatToSlashedFrenchDate } from 'libs/dates'
 import { formatToHour } from 'libs/parsers/formatDates'
 import { BicolorClock } from 'ui/svg/icons/BicolorClock'
