+++ src/features/profile/helpers/computeCredit.ts
@@ -1,0 +1,5 @@
+import { DomainsCredit } from 'api/gen/api'
+
+export const computeCredit = (domainsCredit?: DomainsCredit | null) => {
+  return domainsCredit ? domainsCredit.all.remaining : 0
+}
