+++ src/features/profile/helpers/isUserBeneficiary.ts
@@ -1,0 +1,5 @@
+import { UserProfileResponse } from 'api/gen/api'
+
+export function isUserBeneficiary(user: UserProfileResponse): boolean {
+  return user.isBeneficiary
+}
