+++ src/features/profile/helpers/isUserUnderage.ts
@@ -1,0 +1,4 @@
+import { UserProfileResponse, EligibilityType } from 'api/gen/api'
+
+export const isUserUnderage = (user?: UserProfileResponse) =>
+  user?.eligibility === EligibilityType.underage
