+++ src/features/profile/helpers/isUserUnderageBeneficiary.ts
@@ -1,0 +1,7 @@
+import { UserProfileResponse, UserRole } from 'api/gen/api'
+
+export function isUserUnderageBeneficiary(user: UserProfileResponse | undefined): boolean {
+  if (!user) return false
+  const hasUserUnderageRole = user.roles?.find((role) => role === UserRole.UNDERAGE_BENEFICIARY)
+  return !!hasUserUnderageRole
+}
