+++ src/features/profile/helpers/useIsUserUnderage.ts
@@ -1,0 +1,8 @@
+import { useAuthContext } from 'features/auth/AuthContext'
+
+import { isUserUnderage } from './isUserUnderage'
+
+export const useIsUserUnderage = () => {
+  const { user } = useAuthContext()
+  return isUserUnderage(user)
+}
