+++ src/features/profile/helpers/useIsUserUnderageBeneficiary.ts
@@ -1,0 +1,8 @@
+import { useAuthContext } from 'features/auth/AuthContext'
+
+import { isUserUnderageBeneficiary } from './isUserUnderageBeneficiary'
+
+export const useIsUserUnderageBeneficiary = () => {
+  const { user } = useAuthContext()
+  return isUserUnderageBeneficiary(user)
+}
