+++ src/features/search/helpers/useMaxPrice/useMaxPrice.test.ts
@@ -1,7 +1,7 @@
 import { mocked } from 'ts-jest/utils'
 
 import { useAuthContext } from 'features/auth/AuthContext'
-import { isUserExBeneficiary } from 'features/profile/utils'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { useMaxPrice } from 'features/search/helpers/useMaxPrice/useMaxPrice'
 import { beneficiaryUser, nonBeneficiaryUser } from 'fixtures/user'
 import { renderHook } from 'tests/utils'
