+++ src/features/search/helpers/useMaxPrice/useMaxPrice.ts
@@ -1,5 +1,5 @@
 import { useAuthContext } from 'features/auth/AuthContext'
-import { isUserExBeneficiary } from 'features/profile/utils'
+import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiary'
 import { MAX_PRICE } from 'features/search/helpers/reducer.helpers'
 import { convertCentsToEuros } from 'libs/parsers/pricesConversion'
